requirements.txt
anthropic>=0.25.0
PyGithub>=2.3.0
PyYAML>=6.0
python-dotenv>=1.0.0
rich>=13.7.0
jinja2>=3.1.0
requests>=2.31.0
